requirements.txt
gymnasium[classic-control]
stable-baselines3
torch
groq
sentence-transformers
umap-learn
matplotlib
seaborn
streamlit
numpy
pandas
pyarrow
pydantic
scikit-learn
imageio
imageio-ffmpeg
python-dotenv
tqdm